Create List Creator Part 1
In this program we have a menu displayed in the do...while giving the user the ability to choose from a list of choices to be performed. With in the code of the do...while there are for loops and nested for loops with if and else statements. While the list is empty the the program will tell the user that the list is empty when they choose an option that needs calculations, finding the largest or smallest number and to view the list contents. Any selection given that is not an option displayed will push a notification to the user and have them put in a new entry. To leave, the user must enter a 'q' or 'Q' and the program will exit.