Create List Creator Part 2
This is the updated List Creator that has now implemented a 'SWITCH' statement to allow for easier readability and understanding of what section accesses which function. Functions were added and the for loops along with the nested for loops were moved into the functions for better understanding and readability. The menu options were placed in their own function as well along with a print_list() function that displays the amount of numbers in the list.